Yet another API rewrite
The interaction result (cancel or finished) is now handled at the client level. The result now provides the start and end event and their timing. I found that we need a central place to handle the result and we often need some metadata associated to the last event.

Fixes #40
Fixes #39
Fixes #41